feat: migrate to angular standalone components assemblage (#1119)
BREAKING CHANGE
