+++ .eslintrc.json
@@ -1,9 +1,9 @@
 {
   "root": true,
-  "plugins": ["unused-imports"],
   "overrides": [
     {
       "files": ["*.ts"],
+      "plugins": ["unused-imports"],
       "extends": [
         "plugin:@angular-eslint/recommended",
         "plugin:@angular-eslint/template/process-inline-templates"
