+++ src/app/app.component.html
@@ -11,8 +11,8 @@
         ? 'portal-with-header-footer'
         : 'portal-with-header'
       : hasFooter
-      ? 'portal-with-footer'
-      : 'portal'
+        ? 'portal-with-footer'
+        : 'portal'
   "
   igoStopDropPropagation
 ></app-portal>
