+++ src/app/pages/portal/expansion-panel/expansion-panel-header.component.ts
@@ -7,11 +7,15 @@ import {
   Output
 } from '@angular/core';
 
+import { ExpansionPanelButtonComponent } from './expansion-panel-button/expansion-panel-button.component';
+
 @Component({
   selector: 'app-expansion-panel-header',
   templateUrl: './expansion-panel-header.component.html',
   styleUrls: ['./expansion-panel-header.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [ExpansionPanelButtonComponent]
 })
 export class ExpansionPanelHeaderComponent {
   @Input() expanded: boolean;
