+++ src/app/pages/portal/map-overlay/map-overlay.component.ts
@@ -1,3 +1,4 @@
+import { NgClass, NgFor, NgIf, NgStyle } from '@angular/common';
 import { AfterViewInit, Component, OnDestroy } from '@angular/core';
 
 import { Context, ContextService } from '@igo2/context';
@@ -10,7 +11,9 @@ import { MapOverlay } from './map-overlay.interface';
 @Component({
   selector: 'app-map-overlay',
   templateUrl: './map-overlay.component.html',
-  styleUrls: ['./map-overlay.component.scss']
+  styleUrls: ['./map-overlay.component.scss'],
+  standalone: true,
+  imports: [NgFor, NgClass, NgIf, NgStyle]
 })
 export class MapOverlayComponent implements AfterViewInit, OnDestroy {
   public mapOverlay: MapOverlay[] = [];
