--- src/app/pages/portal/map-overlay/map-overlay.module.ts
@@ -1,11 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-
-import { MapOverlayComponent } from './map-overlay.component';
-
-@NgModule({
-  imports: [CommonModule],
-  exports: [MapOverlayComponent],
-  declarations: [MapOverlayComponent]
-})
-export class MapOverlayModule {}
