--- src/app/pages/portal/toast-panel-for-expansion/toast-panel-for-expansion.module.ts
@@ -1,23 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-
-import { IgoPanelModule, IgoStopPropagationModule } from '@igo2/common';
-import { IgoLanguageModule } from '@igo2/core';
-
-import { ToastPanelForExpansionComponent } from './toast-panel-for-expansion.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    MatIconModule,
-    MatButtonModule,
-    IgoLanguageModule,
-    IgoPanelModule,
-    IgoStopPropagationModule
-  ],
-  exports: [ToastPanelForExpansionComponent],
-  declarations: [ToastPanelForExpansionComponent]
-})
-export class AppToastPanelForExpansionModule {}
