--- src/app/pages/portal/toast-panel/toast-panel.module.ts
@@ -1,37 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { MatBadgeModule } from '@angular/material/badge';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatTooltipModule } from '@angular/material/tooltip';
-
-import {
-  IgoActionModule,
-  IgoPanelModule,
-  IgoStopPropagationModule
-} from '@igo2/common';
-import { IgoLanguageModule } from '@igo2/core';
-import { IgoFeatureModule, IgoSearchResultsModule } from '@igo2/geo';
-
-import { ToastPanelComponent } from './toast-panel.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    MatBadgeModule,
-    MatIconModule,
-    MatButtonModule,
-    MatTooltipModule,
-    MatMenuModule,
-    IgoLanguageModule,
-    IgoPanelModule,
-    IgoStopPropagationModule,
-    IgoActionModule,
-    IgoFeatureModule,
-    IgoSearchResultsModule
-  ],
-  exports: [ToastPanelComponent],
-  declarations: [ToastPanelComponent]
-})
-export class AppToastPanelModule {}
