+++ src/app/pages/portal/welcome-window/welcome-window.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { MatDialogConfig } from '@angular/material/dialog';
 
 import { ConfigService, StorageService } from '@igo2/core';
+
 import { getAppVersion } from 'src/app/app.utils';
 
 @Injectable({
