+++ src/environments/environnement.interface.ts
@@ -1,5 +1,6 @@
 import { AllEnvironmentOptions } from '@igo2/integration';
 import { EnvironmentOptions as IntegrationEnvironmentOptions } from '@igo2/integration';
+
 import { MapOverlay } from 'src/app/pages/portal/map-overlay/map-overlay.interface';
 
 export interface AppEnvironmentOptions extends IntegrationEnvironmentOptions {
